+++ app/src/main/java/com/example/androiddevchallenge/ui/MainActivity.kt
@@ -24,12 +24,14 @@
 package com.example.androiddevchallenge.ui
 
 import android.os.Bundle
+import android.view.View
 import androidx.activity.compose.setContent
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.app.AppCompatDelegate
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.Surface
 import androidx.compose.runtime.Composable
+import androidx.core.splashscreen.SplashScreen
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import com.example.androiddevchallenge.ui.theme.WoofTheme
 import dagger.hilt.android.AndroidEntryPoint
@@ -39,35 +41,17 @@ class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         installSplashScreen()
-
         setContent {
-            val darkTheme = isSystemInDarkTheme()
-            AppContent(darkTheme)
+            AppContent()
         }
     }
 
     @Composable
-    fun AppContent(darkTheme: Boolean) {
+    fun AppContent() {
         WoofTheme {
             Surface {
-                AdoptyApp(
-                    onBulbClicked = {
-                        if (darkTheme) {
-                            setLightTheme()
-                        } else {
-                            setDarkTheme()
-                        }
-                    }
-                )
+                AdoptyApp()
             }
         }
-    }
-
-    private fun setLightTheme() {
-        AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO)
-    }
-
-    private fun setDarkTheme() {
-        AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES)
     }
 }
